+++ doc/CHANGES.rst
@@ -20,6 +20,7 @@ CHANGELOG
 * Fix migration with added more than one new required columns with default values
 * Fix Sql View can define Many2One relationship
 * Fix SQL View, The __mapper__ is put in the Model
+* FIX SQL View, For Sql view the mapping of the field must be find with anyblok prefix column
 
 0.19.1 (2018-06-07)
 -------------------